+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [1.0.0] - 2022-08-27
+### Changed
+* Removed dispatcher invokator and Binding usage to be independent from WPF
+* Rebrand to Chapter.Collections
+
 ## [2.0.0] - 2022-07-23
 ### Changed
 * Upgrade to net6